Testing addition of clickable buttons within app.vue
@heywooddogwood, I have a working snackbar!

The key is that there is a variable within app.vue that sets the state of the click actions, and you have to pass that variable over from app.py (lines 25-29).

Within app.vue, there are values assigned to "v-model" (for example, state.col_tab_model).  Search for those tags for examples of how they are being used.