Histogram/scatter sync improvements.
Intercept the xrange selector in the histogram viewer to set the edit
subset to be the subset that we want to modify. Normally, a glue user
could do this themselves, but the students here don't have access to
the glue UI. Set the HistogramListener to be off when the xrange selector
isn't active, since it otherwise has no purpose.