Match dialogs on page 1 to match mock-up
Clean up the flow of reflection questions at the end of the scaffold sequence to align with the mockup. My version had gotten pretty out of date. Their was a calculate card to randomly generate velocity values that was no longer relevant, as well as two separate reflection sections that should all be appearing together, prior to the mathjax velocity calculation section.
Also, the scaffold burst formatting has been unified a bit, and they have been expanded to better match the mockup as well.