Add header to WWT/Galaxy Picker
Header allows a simpler identifier when referencing in scaffold bursts.

Also there is an issue that the WWT widget gets pushed below the lower margin of the div frame - but I added overflow: hidden; to the style section, which keeps it contained. I don't think this should have any negative impact on the galaxy selection or angular size measurements when implemented there.