Create vue component for scaffold burst
Generic scaffold burst component created that can be used for almost all bursts. Not ALL all, because there are some nuances with the back/next button that don't always work well for generalization. Take the select a galaxy burst for instance. Maybe down the road it can be worked in. For now, this component works for 4/5 of the bursts.