Create intro slideshow shell
Needed to create a new state variable to hold whether we are in the intro section or in the main section. Called it state.stage. Stuffed an element for the intro slideshow into the app.vue code. Created a new vue_component called intro_windows.vue to hold code for this element.

Also tidied up code for specvel and reflect_velocity a little bit.