add image files, update dialogs
Have added a selection of spectrum image / spectrum graph images that will populate the introductory slideshow. These may need to be updated with time, but they will allow us to set formatting and layout. Also have begun updating the design of the slideshow according to the mockup on google drive.