Update grid style
We've got a new layout here, in line with Alyssa's suggestion from the first couple slides of the mockup google slides file (at least, they are the first couple slides currently). The new layout puts the bursts in a right hand column that move up and down with the relevant widget to focus on at any given point. Functionality regarding flow from widget to widget and step to step is maintained as before. Strictly cosmetic.

Although I did accidentally add an additional burst to split up the Select 5 Galaxies burst. I'll include that in the main as well.