Remove "next" button from the "Choose a Row" burst
Pretty self-explanatory. This did require transforming the burst into a standard v-alert as opposed to the scaffold-alert component built specific to this program. Now the only thing that moves the sequence along is clicking on one of the rows.

At this point when you navigate backward through the bursts, the spectrum tool disappears. I think we'll probably want to keep it active once it's been added to the screen.