intro slideshow, fix text above wwt widget
We want the height of the text space above the WWT widget to be a consistent height so that the WWT widget appears in the same vertical location across slides 3-4-5. This has been done by fixing the height of the div above the widget. However, it only sets a min-height, so if the window width contracts enough, all text divs will exceed this min anyway and they will no longer maintain consistent height.