Suppress unnecessary warnings
The Spectrum Viewer Tool uses a glupyter component that expects a more complex data set than we are using, so it is complaining about things it can't find, even though that does not affect the functionality for our purposes.

Beware that we may inadvertently filter out warnings that are actually useful, so remember to comment these lines out if we're having trouble sorting something else out later.