Lighten 2ndary color for light theme
I changed this because the blue cards on several of the slides were not readable in light mode, but now it's making the viewer title bars look bad because those still have white text.

I'm going to leave it for now and let @heywooddogwood continue ironing out the light theme colors.