Small things
Add a variable to accommodate clicking the lambda button in the spectrum viewer. Also create a new guidance element for rest wavelength instructions, however this hasn't been implemented yet and we will continue to use the existing alert. Finally, add a white background to the viewers so that we don't have to do any fancy formatting to the tickmarks in dark/light.