nice work NEXT depends on REFLECT completion
Changed out nice_work_alert for nice_work_guidance so that we can add dependencies for the "next" button. Now we have a new state variable called spec_reflect_complete in stage_one.py that listens for submission of the reflection dialog. At that point, NEXT button will appear in the scaffold.

Also updated the reset view button to mdi-cached in the toolbar for spectrum tool, and in the scaffold for the galaxy selection.