add negative margin to highlighted widgets
When widgets get highlighted, adding the border highlight crams the widget in ever so slightly. I couldn't figure out how to expand the x-directional margin into negative space to accommodate this. However, y-direction is working, so now it doesn't look as awkward when the highlight kicks in.

Also, I deleted the stray select galaxies guidance in the vue_components folder that didn't seem to be doing anything anymore.