restwave guideline NEXT after lambda click
The NEXT button in the rest wavelength guideline is now dependent upon the lambda button in the toolbar of the spectrum tool having been clicked. This is done by creating a callback listener in stage_one.py, which updates the stage_state lambda_used variable when it observes that the lambda_used variable in the rest_wavelength_tool has been updated.