return dialog/card themes to light/dark
We had tried making all the cards/dialogs white backgrounded, so that the several white images wouldn't have a sharp, distracting border. Ultimately we opted against it so that the cards/dialogs would follow the same light/dark theme variations that the rest of the app does. Instead, placed a lot of the white background images into container cards with white backgrounds and rounded edges, so that the border edge is softened.

Also turned off the FIND VELOCITY button for now, since it doesn't seem to be useful in this app. Perhaps can be removed entirely.