Add new thresholds for including a decimal point
It felt weird to me that the starting galaxy field of view is 90" but without the decimal, it's presented as 1' for a long time. For double digit numbers, the decimal felt unnecessary and made the string too long.

Thanks to Jon for helping to figure out the right limits to avoid weird edge cases.