prevent galaxies in table from being re-selected
Needed to turn of the (+) button that appears when a galaxy that has already been added to the My Galaxies table is selected. This entailed adding a candidate_galaxy variable to the selection_tool.py file, in addition to the current_galaxy variable that existed. Both variables are updated when a galaxy is selected, but if it is already in the table, the candidate_galaxy variable is emptied. The button listens for whether the candidate_galaxy variable is empty or not.