Add scroll action to all guidelines
Jon found the appropriate code, I have copied and pasted into each of the guideline elements from what he posted on slack. Whenever element should newly appear on screen (whether for the first time or navigating back/forward), the screen will scroll to make sure the element is in the FOV.

Also fixed the coloring of the footer as a little bonus.