return to full sky view when rows are deselected
Whenever selected rows/galaxies get deselected (such as when we move on to the Choose a Row guideline, or to the doppler calculator, have the cosmic sky viewer reset to original zoom/centering.

Also, I have renamed a few other of the guidelines, although the content of the elements is nearly 100% unchanged.