experiment with making all buttons green
we talked about it being the same green as the FOV marker. But this is also the green of the highlighted galaxies in the selection tool, and it starts to get a bit crowded with so much monochromatic green. Tried using a different shade of green. We'll need to discuss if it works.

Also deleted the velocity calculating button from stage one since it is no longer relevant.