add countdown to ang_size measurements
Copied roughly from the code included to coutndown the observed wavelength measurements in stage 1 / spectrum tool. When a measurement is being made, checks whether or not there is an existing measurement for the angular size for the indexed galaxy in the table. If there isn't, the state variable counting measurements angsize_total increments by 1. Once we hit 5, the next button appears for the guideline.