fix disappearing guideline issue
There's a snag in stage 1 after a user selects their first galaxy and it introduces the My Galaxy table. If they select another galaxy without first clicking next on the table guideline, the guidelines disappear. Needed to add a condition to update_count for stage_one that updates the marker if a galaxy is selected early. And dump the conditions dependent on a galaxy being actively selected.