Build 2 column / first 3 guidelines
We use 2 column system for stage 3: Explore Data again. Only difference is that column breakdown is 6-6, not 4-8.

Also create the first 3 guidelines for the learning experience. These are effectively placeholders. They take from the basic content/design of the mockup, but the mockup will be updated for more effective language / functionality.