remove title header
Tab headers at the top of any page is taking up potentially unnecessary space, so we're cutting it out. The item still exists, but the display has been set to none.

Also, add a couple styling properties to fix background/text color for the toggle card when exploring personal/class data on stage 3.