change layer-toggle background to light gray
Had been a dark-ish blue-grey. However, I think it ought to be a light color to match the fact that the plot point colors are all chosen to offset the white of the plot background. If we have a dark background for the toggle, it gets too complicated trying to get everything to work.
Set the data set title to bold to make it stand out more too.